cli/requirements.txt
# Witmind CLI dependencies
click>=8.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
pyyaml>=6.0
rich>=13.0.0
